requirements.txt
pillow
qrcode
opencv-python
